--- src/BlazorRx.Core/Store/IStore.cs
@@ -1,8 +1,0 @@
-﻿using System;
-namespace BlazorRx.Core.Store
-{
-    public interface IStore<T> where T : class
-    {
-        T GetCurrentState();
-    }
-}
