+++ src/Core/BlazorRx.Core.csproj
