+++ src/Core/Builder/IStoreBuilder.cs
