+++ src/Core/Reducer/IReducer.cs
