+++ test/Core/Model/SimpleClass.cs
