+++ test/Core/Model/SimpleClassSubset.cs
