+++ test/Core/Model/SimpleClassUtilities.cs
