+++ test/Core/Store/StoreTests.cs
